+++ src/crypt/web-crypto.ts
@@ -6,17 +6,19 @@
  */
 
 import type { InstanceConfig } from '../instance/instance.js';
-import type { CryptModule } from './types.js';
+import type { CryptFnContext, CryptModule } from './types.js';
+
+async function crypt(this: 'decrypt' | 'encrypt', { encAlg, key, nonce, payload }: CryptFnContext) {
+  return new Uint8Array(await crypto.subtle[this]({ name: encAlg, iv: nonce }, key, payload));
+}
 
 /**
  * An {@link CryptModule} that implements encryption using the WebCrypto API for supported
  * algorithms.
  */
 export const WebCrypto: CryptModule = {
-  decrypt: async ({ encAlg, key, nonce, payload }) =>
-    crypto.subtle.decrypt({ name: encAlg, iv: nonce }, key, payload),
-  encrypt: async ({ encAlg, key, nonce, payload }) =>
-    crypto.subtle.encrypt({ name: encAlg, iv: nonce }, key, payload),
+  decrypt: crypt.bind('decrypt'),
+  encrypt: crypt.bind('encrypt'),
 };
 
 /**
@@ -24,7 +26,15 @@ export const WebCrypto: CryptModule = {
  * API.
  */
 export const WithWebCrypto: InstanceConfig = {
-  cryptAlgs: {
-    'AES-GCM': WebCrypto,
-  },
+  cryptAlgs: Object.fromEntries(
+    [
+      //
+      'AES-CBC',
+      'AES-CTR',
+      'AES-GCM',
+      'AES-OCB',
+      'ChaCha20-Poly1305',
+      'RSA-OAEP',
+    ].map((alg) => [alg, WebCrypto] as const),
+  ),
 };